move kimi-linear PD parity mixin into server_fixtures
